+++ src/render/api_layer/src/rds_render_api_layer/thread/rdsRenderFrameParam.h
@@ -25,7 +25,7 @@ public:
 public:
 	u64 engineFrameCount()	const;
 	u64 frameCount()		const;
-	u64 frameIndex()		const;
+	u32 frameIndex()		const;
 
 	/*
 	* only use in engine thread
